+++ sdk/metric/instrument.go
@@ -24,7 +24,7 @@ import (
 	"go.opentelemetry.io/otel/metric/embedded"
 	"go.opentelemetry.io/otel/sdk/instrumentation"
 	"go.opentelemetry.io/otel/sdk/metric/aggregation"
-	"go.opentelemetry.io/otel/sdk/metric/internal"
+	"go.opentelemetry.io/otel/sdk/metric/internal/aggregate"
 	"go.opentelemetry.io/otel/sdk/metric/metricdata"
 )
 
@@ -145,8 +145,31 @@ type Stream struct {
 	Unit string
 	// Aggregation the stream uses for an instrument.
 	Aggregation aggregation.Aggregation
-	// AttributeFilter applied to all attributes recorded for an instrument.
-	AttributeFilter attribute.Filter
+	// AllowAttributeKeys are an allow-list of attribute keys that will be
+	// preserved for the stream. Any attribute recorded for the stream with a
+	// key not in this slice will be dropped.
+	//
+	// If this slice is empty, all attributes will be kept.
+	AllowAttributeKeys []attribute.Key
+}
+
+// attributeFilter returns an attribute.Filter that only allows attributes
+// with keys in s.AttributeKeys.
+//
+// If s.AttributeKeys is empty an accept-all filter is returned.
+func (s Stream) attributeFilter() attribute.Filter {
+	if len(s.AllowAttributeKeys) <= 0 {
+		return func(kv attribute.KeyValue) bool { return true }
+	}
+
+	allowed := make(map[attribute.Key]struct{})
+	for _, k := range s.AllowAttributeKeys {
+		allowed[k] = struct{}{}
+	}
+	return func(kv attribute.KeyValue) bool {
+		_, ok := allowed[kv.Key]
+		return ok
+	}
 }
 
 // streamID are the identifying properties of a stream.
@@ -171,7 +194,7 @@ type streamID struct {
 }
 
 type int64Inst struct {
-	aggregators []internal.Aggregator[int64]
+	aggregators []aggregate.Aggregator[int64]
 
 	embedded.Int64Counter
 	embedded.Int64UpDownCounter
@@ -192,7 +215,7 @@ func (i *int64Inst) Record(ctx context.Context, val int64, opts ...metric.Record
 	i.aggregate(ctx, val, c.Attributes())
 }
 
-func (i *int64Inst) aggregate(ctx context.Context, val int64, s attribute.Set) {
+func (i *int64Inst) aggregate(ctx context.Context, val int64, s attribute.Set) { // nolint:revive  // okay to shadow pkg with method.
 	if err := ctx.Err(); err != nil {
 		return
 	}
@@ -202,7 +225,7 @@ func (i *int64Inst) aggregate(ctx context.Context, val int64, s attribute.Set) {
 }
 
 type float64Inst struct {
-	aggregators []internal.Aggregator[float64]
+	aggregators []aggregate.Aggregator[float64]
 
 	embedded.Float64Counter
 	embedded.Float64UpDownCounter
@@ -254,7 +277,7 @@ var _ metric.Float64ObservableCounter = float64Observable{}
 var _ metric.Float64ObservableUpDownCounter = float64Observable{}
 var _ metric.Float64ObservableGauge = float64Observable{}
 
-func newFloat64Observable(scope instrumentation.Scope, kind InstrumentKind, name, desc, u string, agg []internal.Aggregator[float64]) float64Observable {
+func newFloat64Observable(scope instrumentation.Scope, kind InstrumentKind, name, desc, u string, agg []aggregate.Aggregator[float64]) float64Observable {
 	return float64Observable{
 		observable: newObservable(scope, kind, name, desc, u, agg),
 	}
@@ -273,7 +296,7 @@ var _ metric.Int64ObservableCounter = int64Observable{}
 var _ metric.Int64ObservableUpDownCounter = int64Observable{}
 var _ metric.Int64ObservableGauge = int64Observable{}
 
-func newInt64Observable(scope instrumentation.Scope, kind InstrumentKind, name, desc, u string, agg []internal.Aggregator[int64]) int64Observable {
+func newInt64Observable(scope instrumentation.Scope, kind InstrumentKind, name, desc, u string, agg []aggregate.Aggregator[int64]) int64Observable {
 	return int64Observable{
 		observable: newObservable(scope, kind, name, desc, u, agg),
 	}
@@ -283,10 +306,10 @@ type observable[N int64 | float64] struct {
 	metric.Observable
 	observablID[N]
 
-	aggregators []internal.Aggregator[N]
+	aggregators []aggregate.Aggregator[N]
 }
 
-func newObservable[N int64 | float64](scope instrumentation.Scope, kind InstrumentKind, name, desc, u string, agg []internal.Aggregator[N]) *observable[N] {
+func newObservable[N int64 | float64](scope instrumentation.Scope, kind InstrumentKind, name, desc, u string, agg []aggregate.Aggregator[N]) *observable[N] {
 	return &observable[N]{
 		observablID: observablID[N]{
 			name:        name,